Enable undo/redo for DeleteKey

+++ src/main/java/org/jusecase/properties/usecases/DeleteKey.java
@@ -1,10 +1,13 @@
 package org.jusecase.properties.usecases;
 
+import org.jusecase.Usecase;
+import org.jusecase.properties.entities.Property;
+import org.jusecase.properties.entities.UndoableRequest;
+import org.jusecase.properties.gateways.PropertiesGateway;
+
 import javax.inject.Inject;
 import javax.inject.Singleton;
-
-import org.jusecase.Usecase;
-import org.jusecase.properties.gateways.PropertiesGateway;
+import java.util.List;
 
 
 @Singleton
@@ -19,12 +22,22 @@ public class DeleteKey implements Usecase<DeleteKey.Request, DeleteKey.Response>
 
     @Override
     public Response execute(Request request) {
-        propertiesGateway.deleteKey(request.key);
+        if (request.undo) {
+            propertiesGateway.addProperties(request.deletedProperties);
+        } else {
+            request.deletedProperties = propertiesGateway.getProperties(request.key);
+            propertiesGateway.deleteKey(request.key);
+        }
         return new Response();
     }
 
-    public static class Request {
+    public static class Request extends UndoableRequest {
         public String key;
+        List<Property> deletedProperties;
+
+        public Request() {
+            name = "delete key";
+        }
     }
 
     public static class Response {
